requirements.txt
streamlit
python-jobspy
pandas
requests
openai
